// sample-ui/src/api/interfaces/UserApplication.ts
/* tslint:disable */

/**
 * This file was automatically generated by "Swaxios".
 * It should not be modified by hand.
 */

import {AppRoleAssignment} from './AppRoleAssignment';

export interface UserApplication {
  appId?: string;
  resourceId?: string;
  userAssignedAppRoles?: Array<AppRoleAssignment>;
  displayName?: string;
}