// sample-ui/src/api/interfaces/AppRole.ts
/* tslint:disable */

/**
 * This file was automatically generated by "Swaxios".
 * It should not be modified by hand.
 */

export interface AppRole {
  allowedMemberTypes?: Array<string>;
  description?: string;
  displayName?: string;
  id?: string;
  isEnabled?: boolean;
  origin?: string;
  value?: string;
  _odata_type?: string;
}